Add tool to convert XLSX files to loin XML
This tool relies on the structure of the Excel file to be as expected.
Given the provided example, a LOIN XML file is created with one
Specification for each phase of the project, which contains the
Properties that are required for each part in that phase.
As the Excel file (and probably all of those and similar files) does not
contain any information about units, or datatypes, this information
cannot be constained in the result.